requirements.txt
GitPython==1.0.0
